requirements.txt
aiohttp
markdownify
bs4
inquirer
vmd
selenium
readabilipy
python-dotenv